fixed parsing

+++ headers/stddef.h
@@ -1,2 +1,2 @@
-typedef unsigned size_t;
+typedef long unsigned int size_t;
 #define NULL ((void *)0)
